pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skill-gen"
version = "0.1.0"
description = "AI-powered skill generator that uses browser-use to research tools and synthesize Claude Code skills"
readme = "SKILL.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "skill-gen contributors" },
]
keywords = ["ai", "claude", "skills", "browser-automation", "code-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "browser-use>=0.1.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-openai>=0.1.0",
]

[project.optional-dependencies]
google = [
    "langchain-google-genai>=1.0.0",
]
all-llms = [
    "langchain-anthropic>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-google-genai>=1.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[project.scripts]
skill-gen = "skill_gen.cli:main"

[project.urls]
Homepage = "https://github.com/skill-gen/skill-gen"
Repository = "https://github.com/skill-gen/skill-gen"
Issues = "https://github.com/skill-gen/skill-gen/issues"

[tool.hatch.build.targets.wheel]
packages = ["skill_gen"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "SIM108", # use ternary operator instead of if-else
]

[tool.ruff.lint.isort]
known-first-party = ["skill_gen"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring browser-use and network access",
]
